requirements.txt
asgiref==3.11.1
attrs==25.4.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
cryptography==46.0.5
defusedxml==0.7.1
Django==3.2.16
django-filter==2.4.0
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
djoser==2.1.0
idna==3.11
iniconfig==2.3.0
itypes==1.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
oauthlib==3.3.1
packaging==26.0
Pillow==9.3.0
pluggy==1.0.0.dev0
py==1.11.0
pycparser==3.0
PyJWT==2.11.0
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
python3-openid==3.2.0
pytz==2026.1.post1
requests==2.32.5
requests-oauthlib==2.0.0
six==1.17.0
social-auth-app-django==4.0.0
social-auth-core==4.8.5
sqlparse==0.5.5
toml==0.10.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==1.26.20
